# api-docs-app.json
{"openapi":"3.0.1","info":{"title":"å¥èº«å¹³å°ç¨æ·ç«¯API","description":"å¥èº«å¹³å°ç¨æ·ç«¯APPæ¥å£ - Spring Boot 3.0.5 + MyBatis-Plus 3.5.3.1","termsOfService":"https://github.com/your-project","license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0"},"version":"1.0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"tags":[{"name":"Appç«¯ä¸ªäººä¿¡æ¯ç®¡ç","description":"ä¸ªäººèµææ¥çãä¿®æ¹ãå¤´åä¸ä¼ ãå¯ç ä¿®æ¹ãè´¦å·æ³¨éç­åè½"},{"name":"Appç«¯è®¤è¯ç®¡ç","description":"ç¨æ·æ³¨åãç»å½ãéªè¯ç ç­åè½"}],"paths":{"/app/profile/password":{"put":{"tags":["Appç«¯ä¸ªäººä¿¡æ¯ç®¡ç"],"summary":"ä¿®æ¹å¯ç ","description":"ä¿®æ¹ç»å½å¯ç ","operationId":"updatePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordUpdateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultString"}}}}}}},"/app/profile/info":{"get":{"tags":["Appç«¯ä¸ªäººä¿¡æ¯ç®¡ç"],"summary":"è·åä¸ªäººä¿¡æ¯","description":"è·åå½åç»å½ç¨æ·çè¯¦ç»ä¿¡æ¯","operationId":"getProfileInfo","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultProfileDetailVO"}}}}}},"put":{"tags":["Appç«¯ä¸ªäººä¿¡æ¯ç®¡ç"],"summary":"ä¿®æ¹ä¸ªäººä¿¡æ¯","description":"æ´æ°ä¸ªäººåºæ¬ä¿¡æ¯","operationId":"updateProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileUpdateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultString"}}}}}}},"/app/profile/cancel":{"post":{"tags":["Appç«¯ä¸ªäººä¿¡æ¯ç®¡ç"],"summary":"ç³è¯·è´¦å·æ³¨é","description":"ç³è¯·æ³¨éç¨æ·è´¦å·","operationId":"cancelAccount","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountCancelDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultAccountCancelVO"}}}}}}},"/app/profile/cancel/{cancelId}/revoke":{"post":{"tags":["Appç«¯ä¸ªäººä¿¡æ¯ç®¡ç"],"summary":"æ¤éæ³¨éç³è¯·","description":"æ¤éä¹åçè´¦å·æ³¨éç³è¯·","operationId":"revokeCancelApplication","parameters":[{"name":"cancelId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultString"}}}}}}},"/app/profile/avatar/upload":{"post":{"tags":["Appç«¯ä¸ªäººä¿¡æ¯ç®¡ç"],"summary":"ä¸ä¼ å¤´å","description":"ä¸ä¼ ç¨æ·å¤´åå¾ç","operationId":"uploadAvatar","requestBody":{"content":{"application/json":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultAvatarUploadVO"}}}}}}},"/app/auth/sms-code/send":{"post":{"tags":["Appç«¯è®¤è¯ç®¡ç"],"summary":"åéç­ä¿¡éªè¯ç ","description":"åéæ³¨åãç»å½æéç½®å¯ç éªè¯ç ","operationId":"sendSmsCode","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SmsCodeDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultString"}}}}}}},"/app/auth/register":{"post":{"tags":["Appç«¯è®¤è¯ç®¡ç"],"summary":"ç¨æ·æ³¨å","description":"æ°ç¨æ·æ³¨å","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegisterDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultRegisterVO"}}}}}}},"/app/auth/refresh-token":{"post":{"tags":["Appç«¯è®¤è¯ç®¡ç"],"summary":"å·æ°ä»¤ç","description":"ä½¿ç¨refresh tokenè·åæ°çaccess token","operationId":"refreshToken","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultRefreshTokenVO"}}}}}}},"/app/auth/logout":{"post":{"tags":["Appç«¯è®¤è¯ç®¡ç"],"summary":"éåºç»å½","description":"ç¨æ·éåºç»å½","operationId":"logout","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultString"}}}}}}},"/app/auth/login":{"post":{"tags":["Appç«¯è®¤è¯ç®¡ç"],"summary":"ç¨æ·ç»å½","description":"å¯ç ç»å½æç­ä¿¡éªè¯ç ç»å½","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultLoginVO"}}}}}}},"/app/auth/forget-password":{"post":{"tags":["Appç«¯è®¤è¯ç®¡ç"],"summary":"å¿è®°å¯ç ","description":"éè¿ç­ä¿¡éªè¯ç éç½®å¯ç ","operationId":"forgetPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgetPasswordDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultString"}}}}}}},"/app/profile/statistics":{"get":{"tags":["Appç«¯ä¸ªäººä¿¡æ¯ç®¡ç"],"summary":"è·åè´¦å·ç»è®¡ä¿¡æ¯","description":"è·åç¨æ·è´¦å·ç¸å³çç»è®¡æ°æ®","operationId":"getAccountStatistics","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultObject"}}}}}}}},"components":{"schemas":{"PasswordUpdateDTO":{"required":["confirmPassword","newPassword","oldPassword"],"type":"object","properties":{"oldPassword":{"type":"string","description":"æ§å¯ç ","example":"oldPassword123"},"newPassword":{"maxLength":20,"minLength":6,"type":"string","description":"æ°å¯ç ","example":"newPassword123"},"confirmPassword":{"type":"string","description":"ç¡®è®¤æ°å¯ç ","example":"newPassword123"}},"description":"å¯ç ä¿®æ¹è¯·æ±"},"ResultString":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"string"}}},"ProfileUpdateDTO":{"type":"object","properties":{"nickname":{"maxLength":50,"minLength":0,"type":"string","description":"æµç§°","example":"æ°æµç§°"},"email":{"maxLength":100,"minLength":0,"type":"string","description":"é®ç®±","example":"user@example.com"},"gender":{"maximum":2,"minimum":0,"type":"integer","description":"æ§å«ï¼0-æªç¥ï¼1-ç·ï¼2-å¥³","format":"int32","example":1},"birthDate":{"type":"string","description":"åºçæ¥æ","format":"date","example":"1990-01-01"},"bio":{"maxLength":500,"minLength":0,"type":"string","description":"ä¸ªäººç®ä»","example":"è¿æ¯æçä¸ªäººç®ä»"},"address":{"maxLength":200,"minLength":0,"type":"string","description":"å°å","example":"åäº¬å¸æé³åº"},"occupation":{"maxLength":50,"minLength":0,"type":"string","description":"èä¸","example":"è½¯ä»¶å·¥ç¨å¸"},"height":{"maximum":250,"minimum":100,"type":"integer","description":"èº«é«ï¼åç±³ï¼","format":"int32","example":175},"weight":{"maximum":200,"minimum":30,"type":"integer","description":"ä½éï¼å¬æ¤ï¼","format":"int32","example":70}},"description":"ä¸ªäººä¿¡æ¯ä¿®æ¹è¯·æ±"},"AccountCancelDTO":{"required":["cancelType","password","reason","smsCode"],"type":"object","properties":{"reason":{"type":"string","description":"æ³¨éåå ","example":"ä¸åä½¿ç¨"},"cancelType":{"pattern":"^(temporary|permanent)$","type":"string","description":"æ³¨éç±»åï¼temporary-ä¸´æ¶æ³¨éï¼permanent-æ°¸ä¹æ³¨é","example":"temporary"},"password":{"type":"string","description":"å¯ç ç¡®è®¤","example":"password123"},"smsCode":{"pattern":"^\\d{6}$","type":"string","description":"ç­ä¿¡éªè¯ç ï¼å¼åé¶æ®µå¯ä¼ åºå®å¼123456ï¼","example":"123456"},"additionalNote":{"type":"string","description":"éå è¯´æ","example":"å¶ä»è¯¦ç»è¯´æ"}},"description":"è´¦å·æ³¨éè¯·æ±"},"AccountCancelVO":{"type":"object","properties":{"cancelId":{"type":"integer","description":"æ³¨éç³è¯·ID","format":"int64"},"userId":{"type":"integer","description":"ç¨æ·ID","format":"int64"},"cancelType":{"type":"string","description":"æ³¨éç±»åï¼temporary-ä¸´æ¶æ³¨éï¼permanent-æ°¸ä¹æ³¨é"},"status":{"type":"string","description":"æ³¨éç¶æï¼pending-å¾å¤çï¼approved-å·²æ¹åï¼rejected-å·²æç»"},"applyTime":{"type":"string","description":"ç³è¯·æ¶é´","format":"date-time"},"effectiveTime":{"type":"string","description":"é¢è®¡çææ¶é´","format":"date-time"},"reason":{"type":"string","description":"æ³¨éåå "},"message":{"type":"string","description":"å¤çç»ææ¶æ¯"},"cancellable":{"type":"boolean","description":"æ¯å¦å¯æ¤é"},"cancelDeadline":{"type":"string","description":"æ¤éæªæ­¢æ¶é´","format":"date-time"}},"description":"è´¦å·æ³¨éååº"},"ResultAccountCancelVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/AccountCancelVO"}}},"AvatarUploadVO":{"type":"object","properties":{"avatarUrl":{"type":"string","description":"å¤´åè®¿é®URL"},"fileName":{"type":"string","description":"å¤´åæä»¶å"},"fileSize":{"type":"integer","description":"æä»¶å¤§å°ï¼å­èï¼","format":"int64"},"uploadTime":{"type":"string","description":"ä¸ä¼ æ¶é´","format":"date-time"},"status":{"type":"string","description":"ä¸ä¼ ç¶æï¼success-æåï¼failed-å¤±è´¥"},"message":{"type":"string","description":"ç¶ææ¶æ¯"}},"description":"å¤´åä¸ä¼ ååº"},"ResultAvatarUploadVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/AvatarUploadVO"}}},"SmsCodeDTO":{"required":["phone","type"],"type":"object","properties":{"phone":{"pattern":"^1[3-9]\\d{9}$","type":"string","description":"ææºå·","example":"13888888888"},"type":{"pattern":"^(register|login|reset_password)$","type":"string","description":"éªè¯ç ç±»åï¼register-æ³¨åï¼login-ç»å½ï¼reset_password-éç½®å¯ç ","example":"register"},"captcha":{"type":"string","description":"å¾å½¢éªè¯ç ï¼å¦æå¼å¯å¾å½¢éªè¯ç æ ¡éªï¼","example":"abcd"},"captchaKey":{"type":"string","description":"å¾å½¢éªè¯ç keyï¼å¦æå¼å¯å¾å½¢éªè¯ç æ ¡éªï¼","example":"captcha_key_123"}},"description":"ç­ä¿¡éªè¯ç è¯·æ±"},"UserRegisterDTO":{"required":["confirmPassword","nickname","password","phone","smsCode","username"],"type":"object","properties":{"username":{"maxLength":20,"minLength":3,"pattern":"^[a-zA-Z0-9_]+$","type":"string","description":"ç¨æ·å","example":"user123"},"password":{"maxLength":20,"minLength":6,"type":"string","description":"å¯ç ","example":"123456"},"confirmPassword":{"type":"string","description":"ç¡®è®¤å¯ç ","example":"123456"},"nickname":{"maxLength":50,"minLength":0,"type":"string","description":"æµç§°","example":"ç¨æ·æµç§°"},"phone":{"pattern":"^1[3-9]\\d{9}$","type":"string","description":"ææºå·","example":"13888888888"},"smsCode":{"pattern":"^\\d{6}$","type":"string","description":"ç­ä¿¡éªè¯ç ï¼å¼åé¶æ®µå¯ä¼ åºå®å¼123456ï¼","example":"123456"},"email":{"maxLength":100,"minLength":0,"type":"string","description":"é®ç®±","example":"user@example.com"},"gender":{"maximum":2,"minimum":0,"type":"integer","description":"æ§å«ï¼0-æªç¥ï¼1-ç·ï¼2-å¥³","format":"int32","example":1},"birthDate":{"type":"string","description":"åºçæ¥æ","format":"date","example":"1990-01-01"},"inviteCode":{"maxLength":20,"minLength":0,"type":"string","description":"éè¯·ç ","example":"ABC123"}},"description":"ç¨æ·æ³¨åè¯·æ±"},"RegisterVO":{"type":"object","properties":{"userId":{"type":"integer","description":"ç¨æ·ID","format":"int64"},"username":{"type":"string","description":"ç¨æ·å"},"nickname":{"type":"string","description":"æµç§°"},"phone":{"type":"string","description":"ææºå·"},"registerTime":{"type":"string","description":"æ³¨åæ¶é´","format":"date-time"},"autoLogin":{"type":"boolean","description":"æ¯å¦èªå¨ç»å½"},"accessToken":{"type":"string","description":"è®¿é®ä»¤çï¼èªå¨ç»å½æ¶è¿åï¼"},"refreshToken":{"type":"string","description":"å·æ°ä»¤çï¼èªå¨ç»å½æ¶è¿åï¼"},"welcomeMessage":{"type":"string","description":"æ¬¢è¿æ¶æ¯"}},"description":"æ³¨åååº"},"ResultRegisterVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/RegisterVO"}}},"RefreshTokenDTO":{"required":["refreshToken"],"type":"object","properties":{"refreshToken":{"type":"string","description":"å·æ°ä»¤ç","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."},"deviceId":{"type":"string","description":"è®¾å¤æ è¯","example":"device_123456"},"deviceType":{"type":"string","description":"è®¾å¤ç±»åï¼androidãiosãweb","example":"android"}},"description":"å·æ°ä»¤çè¯·æ±"},"RefreshTokenVO":{"type":"object","properties":{"accessToken":{"type":"string","description":"æ°çè®¿é®ä»¤ç"},"refreshToken":{"type":"string","description":"æ°çå·æ°ä»¤çï¼å¦æéè¦è½®æ¢ï¼"},"tokenType":{"type":"string","description":"ä»¤çç±»å","example":"Bearer"},"accessTokenExpire":{"type":"string","description":"è®¿é®ä»¤çè¿ææ¶é´","format":"date-time"},"refreshTokenExpire":{"type":"string","description":"å·æ°ä»¤çè¿ææ¶é´","format":"date-time"},"userInfo":{"$ref":"#/components/schemas/UserInfoVO"}},"description":"å·æ°ä»¤çååº"},"ResultRefreshTokenVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/RefreshTokenVO"}}},"UserInfoVO":{"type":"object","properties":{"id":{"type":"integer","description":"ç¨æ·ID","format":"int64"},"username":{"type":"string","description":"ç¨æ·å"},"nickname":{"type":"string","description":"æµç§°"},"email":{"type":"string","description":"é®ç®±"},"phone":{"type":"string","description":"ææºå·"},"avatar":{"type":"string","description":"å¤´å"},"gender":{"type":"integer","description":"æ§å«ï¼0-æªç¥ï¼1-ç·ï¼2-å¥³","format":"int32"},"birthDate":{"type":"string","description":"åºçæ¥æ","format":"date"},"status":{"type":"integer","description":"ç¶æï¼0-ç¦ç¨ï¼1-æ­£å¸¸","format":"int32"},"createTime":{"type":"string","description":"æ³¨åæ¶é´","format":"date-time"},"lastLoginTime":{"type":"string","description":"æåç»å½æ¶é´","format":"date-time"},"vipLevel":{"type":"integer","description":"VIPç­çº§","format":"int32"},"vipExpireTime":{"type":"string","description":"VIPå°ææ¶é´","format":"date-time"}},"description":"ç¨æ·ä¿¡æ¯"},"UserLoginDTO":{"required":["account","loginType","password"],"type":"object","properties":{"account":{"type":"string","description":"ç»å½è´¦å·ï¼ç¨æ·å/ææºå·ï¼","example":"13888888888"},"password":{"type":"string","description":"å¯ç ","example":"123456"},"loginType":{"pattern":"^(password|sms)$","type":"string","description":"ç»å½ç±»åï¼password-å¯ç ç»å½ï¼sms-ç­ä¿¡ç»å½","example":"password"},"smsCode":{"pattern":"^\\d{6}$","type":"string","description":"ç­ä¿¡éªè¯ç ï¼ç­ä¿¡ç»å½æ¶å¿å¡«ï¼å¼åé¶æ®µå¯ä¼ åºå®å¼123456ï¼","example":"123456"},"deviceId":{"type":"string","description":"è®¾å¤æ è¯","example":"device_123456"},"deviceType":{"pattern":"^(android|ios|web)$","type":"string","description":"è®¾å¤ç±»åï¼androidãiosãweb","example":"android"}},"description":"ç¨æ·ç»å½è¯·æ±"},"LoginVO":{"type":"object","properties":{"accessToken":{"type":"string","description":"è®¿é®ä»¤ç"},"refreshToken":{"type":"string","description":"å·æ°ä»¤ç"},"tokenType":{"type":"string","description":"ä»¤çç±»å","example":"Bearer"},"accessTokenExpire":{"type":"string","description":"è®¿é®ä»¤çè¿ææ¶é´","format":"date-time"},"refreshTokenExpire":{"type":"string","description":"å·æ°ä»¤çè¿ææ¶é´","format":"date-time"},"userInfo":{"$ref":"#/components/schemas/UserInfoVO"},"firstLogin":{"type":"boolean","description":"æ¯å¦é¦æ¬¡ç»å½"},"needCompleteProfile":{"type":"boolean","description":"æ¯å¦éè¦å®åèµæ"}},"description":"ç»å½ååº"},"ResultLoginVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/LoginVO"}}},"ForgetPasswordDTO":{"required":["confirmPassword","newPassword","phone","smsCode"],"type":"object","properties":{"phone":{"pattern":"^1[3-9]\\d{9}$","type":"string","description":"ææºå·","example":"13888888888"},"newPassword":{"maxLength":20,"minLength":6,"type":"string","description":"æ°å¯ç ","example":"123456"},"confirmPassword":{"type":"string","description":"ç¡®è®¤æ°å¯ç ","example":"123456"},"smsCode":{"pattern":"^\\d{6}$","type":"string","description":"ç­ä¿¡éªè¯ç ï¼å¼åé¶æ®µå¯ä¼ åºå®å¼123456ï¼","example":"123456"}},"description":"å¿è®°å¯ç è¯·æ±"},"ResultObject":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"object"}}},"ProfileDetailVO":{"type":"object","properties":{"id":{"type":"integer","description":"ç¨æ·ID","format":"int64"},"username":{"type":"string","description":"ç¨æ·å"},"nickname":{"type":"string","description":"æµç§°"},"email":{"type":"string","description":"é®ç®±"},"phone":{"type":"string","description":"ææºå·"},"avatar":{"type":"string","description":"å¤´å"},"gender":{"type":"integer","description":"æ§å«ï¼0-æªç¥ï¼1-ç·ï¼2-å¥³","format":"int32"},"genderText":{"type":"string","description":"æ§å«æ¾ç¤ºææ¬"},"birthDate":{"type":"string","description":"åºçæ¥æ","format":"date"},"age":{"type":"integer","description":"å¹´é¾","format":"int32"},"bio":{"type":"string","description":"ä¸ªäººç®ä»"},"address":{"type":"string","description":"å°å"},"occupation":{"type":"string","description":"èä¸"},"height":{"type":"integer","description":"èº«é«ï¼åç±³ï¼","format":"int32"},"weight":{"type":"integer","description":"ä½éï¼å¬æ¤ï¼","format":"int32"},"status":{"type":"integer","description":"ç¶æï¼0-ç¦ç¨ï¼1-æ­£å¸¸","format":"int32"},"vipLevel":{"type":"integer","description":"VIPç­çº§","format":"int32"},"vipExpireTime":{"type":"string","description":"VIPå°ææ¶é´","format":"date-time"},"createTime":{"type":"string","description":"æ³¨åæ¶é´","format":"date-time"},"lastLoginTime":{"type":"string","description":"æåç»å½æ¶é´","format":"date-time"},"points":{"type":"integer","description":"ç§¯åä½é¢","format":"int32"},"profileCompleted":{"type":"boolean","description":"æ¯å¦å®åèµæ"},"completionRate":{"type":"integer","description":"èµæå®æåº¦ï¼ç¾åæ¯ï¼","format":"int32"}},"description":"ä¸ªäººä¿¡æ¯è¯¦æ"},"ResultProfileDetailVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/ProfileDetailVO"}}}}}}