# web/web-app/src/main/java/com/huang/web/app/controller/AuthController.java
package com.huang.web.app.controller;

import com.huang.common.result.Result;
import com.huang.common.utils.SmsCodeUtil;
import com.huang.web.app.dto.auth.*;
import com.huang.web.app.vo.auth.*;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import jakarta.validation.Valid;

/**
 * App端认证控制器
 * @author system
 * @since 2025-01-24
 */
@Tag(name = "App端认证管理", description = "用户注册、登录、验证码等功能")
@Slf4j
@RestController
@RequestMapping("/app/auth")
@Validated
public class AuthController {

    @Autowired
    private SmsCodeUtil smsCodeUtil;

    // TODO: 注入UserService等业务服务

    @Operation(summary = "发送短信验证码", description = "发送注册、登录或重置密码验证码")
    @PostMapping("/sms-code/send")
    public Result<String> sendSmsCode(@Valid @RequestBody SmsCodeDTO dto) {
        log.info("发送短信验证码请求: 手机号={}, 类型={}", dto.getPhone(), dto.getType());
        
        // 检查发送频率限制
        if (!smsCodeUtil.canSendSms(dto.getPhone(), dto.getType())) {
            return Result.fail("发送过于频繁，请稍后再试");
        }
        
        // 发送验证码
        String code = smsCodeUtil.sendSmsCode(dto.getPhone(), dto.getType());
        
        if (code != null) {
            // 开发模式下返回验证码
            return Result.ok("验证码发送成功，开发模式验证码: " + code);
        }
        
        return Result.ok("验证码发送成功");
    }

    @Operation(summary = "用户注册", description = "新用户注册")
    @PostMapping("/register")
    public Result<RegisterVO> register(@Valid @RequestBody UserRegisterDTO dto) {
        log.info("用户注册请求: 用户名={}, 手机号={}", dto.getUsername(), dto.getPhone());
        
        // 验证确认密码
        if (!dto.getPassword().equals(dto.getConfirmPassword())) {
            return Result.fail("两次输入的密码不一致");
        }
        
        // 验证短信验证码
        if (!smsCodeUtil.verifySmsCode(dto.getPhone(), dto.getSmsCode(), "register")) {
            return Result.fail("验证码错误或已失效");
        }
        
        // TODO: 实现用户注册逻辑
        // 1. 检查用户名和手机号是否已存在
        // 2. 密码加密
        // 3. 创建用户记录
        // 4. 可选择自动登录
        
        // 临时返回模拟数据
        RegisterVO vo = new RegisterVO();
        vo.setUserId(1001L);
        vo.setUsername(dto.getUsername());
        vo.setNickname(dto.getNickname());
        vo.setPhone(dto.getPhone());
        vo.setAutoLogin(true);
        vo.setWelcomeMessage("注册成功，欢迎加入健身平台！");
        
        log.info("用户注册成功: 用户ID={}, 用户名={}", vo.getUserId(), vo.getUsername());
        return Result.ok(vo);
    }

    @Operation(summary = "用户登录", description = "密码登录或短信验证码登录")
    @PostMapping("/login")
    public Result<LoginVO> login(@Valid @RequestBody UserLoginDTO dto) {
        log.info("用户登录请求: 账号={}, 登录类型={}", dto.getAccount(), dto.getLoginType());
        
        if ("sms".equals(dto.getLoginType())) {
            // 短信登录验证
            if (!smsCodeUtil.verifySmsCode(dto.getAccount(), dto.getSmsCode(), "login")) {
                return Result.fail("验证码错误或已失效");
            }
        }
        
        // TODO: 实现登录逻辑
        // 1. 根据账号查询用户（用户名或手机号）
        // 2. 密码登录时验证密码
        // 3. 检查用户状态
        // 4. 生成JWT令牌
        // 5. 更新最后登录时间
        
        // 临时返回模拟数据
        LoginVO vo = new LoginVO();
        
        // 模拟用户信息
        UserInfoVO userInfo = new UserInfoVO();
        userInfo.setId(1001L);
        userInfo.setUsername("testuser");
        userInfo.setNickname("测试用户");
        userInfo.setPhone("13888888888");
        userInfo.setGender(1);
        userInfo.setStatus(1);
        
        vo.setUserInfo(userInfo);
        vo.setFirstLogin(false);
        vo.setNeedCompleteProfile(false);
        
        // TODO: 生成真实的JWT令牌
        vo.setAccessToken("mock-access-token-for-development");
        vo.setRefreshToken("mock-refresh-token-for-development");
        
        log.info("用户登录成功: 用户ID={}, 用户名={}", userInfo.getId(), userInfo.getUsername());
        return Result.ok(vo);
    }

    @Operation(summary = "刷新令牌", description = "使用refresh token获取新的access token")
    @PostMapping("/refresh-token")
    public Result<RefreshTokenVO> refreshToken(@Valid @RequestBody RefreshTokenDTO dto) {
        log.info("刷新令牌请求: refreshToken前8位={}", 
                dto.getRefreshToken().length() > 8 ? dto.getRefreshToken().substring(0, 8) + "..." : dto.getRefreshToken());
        
        // TODO: 实现令牌刷新逻辑
        // 1. 验证refresh token有效性
        // 2. 生成新的access token
        // 3. 可选择生成新的refresh token（令牌轮换）
        
        // 临时返回模拟数据
        RefreshTokenVO vo = new RefreshTokenVO();
        vo.setAccessToken("new-mock-access-token");
        vo.setRefreshToken("new-mock-refresh-token");
        
        log.info("令牌刷新成功");
        return Result.ok(vo);
    }

    @Operation(summary = "忘记密码", description = "通过短信验证码重置密码")
    @PostMapping("/forget-password")
    public Result<String> forgetPassword(@Valid @RequestBody ForgetPasswordDTO dto) {
        log.info("忘记密码请求: 手机号={}", dto.getPhone());
        
        // 验证确认密码
        if (!dto.getNewPassword().equals(dto.getConfirmPassword())) {
            return Result.fail("两次输入的密码不一致");
        }
        
        // 验证短信验证码
        if (!smsCodeUtil.verifySmsCode(dto.getPhone(), dto.getSmsCode(), "reset_password")) {
            return Result.fail("验证码错误或已失效");
        }
        
        // TODO: 实现密码重置逻辑
        // 1. 根据手机号查询用户
        // 2. 更新密码（需要加密）
        // 3. 清理相关缓存和令牌
        
        log.info("密码重置成功: 手机号={}", dto.getPhone());
        return Result.ok("密码重置成功");
    }

    @Operation(summary = "退出登录", description = "用户退出登录")
    @PostMapping("/logout")
    public Result<String> logout() {
        // TODO: 实现退出登录逻辑
        // 1. 从请求头获取当前用户信息
        // 2. 将当前令牌加入黑名单
        // 3. 清理相关缓存
        
        log.info("用户退出登录成功");
        return Result.ok("退出登录成功");
    }
}